🐛 fix(ssh-manager): 修 askpass 特殊字符截断 + 测试拆分合规
- Windows askpass 脚本改用 setlocal enabledelayedexpansion + echo !PW!,
  避免密码含 &|<>^ 等 cmd 特殊字符时被二次解析截断
- server_view.rs 单元测试拆分到 server_view_tests.rs,符合 AGENTS.md 5.6 规则
